fix compiler bug for bit operations on reg vals
U0 f() {
  I64 i=-1;
  i&=~1;
  "%x\n",i;
}

Used to only return the lower 32 bits because of a compiler bug (used 32
bit regs instead of 64 for some reason, and 32 bit ops clear the upper 32
bits on x86_64), now it doesn't do that. I hope someone didn't rely on
this bug to clear the upper 32 bits quoth Hyrum's law.