Improve sleep/awake accuracy on NT
Waitable timers will only wait for ~700μs consistently when asked to
wait for 1ms when the timer resolution is set to the minimum. Calibrating
the unit size in NtSetTimer might do the trick, but I'm lazy and what I
was doing (calling NtSetTimer again with 0 in WakeCoreUp) was already a
poor imitation of POSIX' timer_settime timer disarm code and didn't look
nice, so I'll roll with calling NtDelayExecution and queueing an APC
which seems official, orthodox and accurate.